+++ R/wb_styles.R
@@ -192,7 +192,7 @@ create_numfmt <- function(numFmtId, formatCode) {
 #' @param extend extend
 #' @param family font family: default "2"
 #' @param i italic
-#' @param name font name: default "Calibri"
+#' @param name font name: default "Aptos Narrow"
 #' @param outline outline
 #' @param scheme font scheme: default "minor"
 #' @param shadow shadow
@@ -220,7 +220,7 @@ create_font <- function(
     extend     = "",
     family     = "2",
     i          = "",
-    name       = "Calibri",
+    name       = "Aptos Narrow",
     outline    = "",
     scheme     = "minor",
     shadow     = "",
@@ -701,7 +701,7 @@ get_cell_styles <- function(wb, sheet, cell) {
 #' # change font (type, size and color) and background
 #' # the old default in openxlsx and openxlsx2 <= 0.3
 #' style3 <- create_dxfs_style(
-#'   font_name = "Calibri",
+#'   font_name = "Aptos Narrow",
 #'   font_size = 11,
 #'   font_color = wb_color(hex = "FF9C0006"),
 #'   bgFill = wb_color(hex = "FFFFC7CE")